Add [POST] /register endpoint
In this commit, I added POST /register endpoint for registering a new user to the database.
This endpoint uses bcrypt to hash the password field of the user input before saving it to the database.
This endpoint also utilizes a validateNewUser middleware for sanitizing input for creating a new user